refactor(group-lifecycle): keep the series predicate out of the browser bundle
The headless boundary test failed by two bytes: 217.002 KiB against a
217 KiB ceiling. The cause was my own placement. `group-activation-status.ts`
carries a comment explaining exactly this hazard — the watermark key registry
lives there rather than beside its computation because "importing them from
the computation drags the whole reading module into every browser bundle
that validates a group" — and putting `isSameGroupActivationSeries` in that
same module turned its type-only layout-identity import into a value import
on the validator path.

So the predicate moves to `is-same-group-activation-series.ts` beside it.
Both server call sites and the test import the new module; the status module
is type-only for the browser again. The dedup is unchanged: still one
implementation, still asked by the writer's two guards and the formation
view.

The ceiling is a maintainer decision and stays where it is. Headless brotli
is back under it and the shared-web bundle budgets pass.
