+++ packages/shared/api/group-lifecycle/activation-status/group-activation-status.ts
@@ -1,4 +1,4 @@
-import { isSameGroupLayoutIdentity, type GroupLayoutIdentity } from '../group-layout-identity.ts';
+import type { GroupLayoutIdentity } from '../group-layout-identity.ts';
 import type { GroupActivationCondition } from './compute-group-activation-condition.ts';
 import type { GroupEvidenceWatermark } from './compute-group-formation-reading.ts';
 
@@ -71,24 +71,3 @@ export const GROUP_ACTIVATION_STATUS_KEYS = [
     'evidenceWatermark',
     'publishedAtEpochMs'
 ] as const satisfies readonly (keyof GroupActivationStatus)[];
-
-/**
- * The causal series a status belongs to (product decision 33). Both halves
- * are needed: a changed basis starts a distinct series, and so does a changed
- * epoch — a hold-landing `reconfigure` from `active` advances the epoch while
- * retaining the accepted layout identity, so a basis comparison alone would
- * carry the spent series' band into the live one.
- */
-export type GroupActivationSeries = Readonly<{
-    formationEpoch: number;
-    coverageBasisLayoutIdentity: GroupLayoutIdentity;
-}>;
-
-/** Whether a stored status still describes `series`. */
-export function isSameGroupActivationSeries(
-    status: GroupActivationSeries,
-    series: GroupActivationSeries
-): boolean {
-    return status.formationEpoch === series.formationEpoch &&
-        isSameGroupLayoutIdentity(status.coverageBasisLayoutIdentity, series.coverageBasisLayoutIdentity);
-}
